mon: MonitorDBStore: override generic bool leveldb options on init
Although the default option for MonitorDBStore's compression was false,
we were only overriding whatever was the default for the generic leveldb
options IFF we happened to set said option to 'true'.  This would then
degenerate in having this option being completely pointless if it was set
to false (which was the default) and having leveldb compressing on the
monitors.

According to 'perf', there was a significant amount of time being spent
compressing blocks during compaction.  After this patch, perf no longer
showed any compression happening.

Hopefully, this will reduce the cpu strain on big clusters.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>